primitive udf support

+++ executor/spark/src/main/scala/sparkutils/loader/GisticLoader.scala
@@ -7,8 +7,8 @@ import scala.collection.mutable.HashMap
 import scala.collection.mutable.ArrayBuffer
 import java.io.File
 
-case class Sample(gistic_sample: String, focal_score: Int)
-case class Gistic(gistic_gene: String, gistic_gene_id: Int, cytoband: String, gistic_samples: Seq[Sample])
+case class Sample(gistic_sample: String, focal_score: Long)
+case class Gistic(gistic_gene: String, gistic_gene_id: Long, cytoband: String, gistic_samples: Seq[Sample])
 //  {(gene: String, cytoband: String, samples: {(name: String, focal_score: Int)})}
 
 class GisticLoader(spark: SparkSession) {
@@ -50,8 +50,8 @@ class GisticLoader(spark: SparkSession) {
     val df = iterMerge(ldf, files.tail)
     val columns = spark.sparkContext.broadcast(df.columns.toSet -- Set("Gene Symbol", "Gene ID", "Cytoband"))
     df.mapPartitions{ it => it.map{ r => 
-        Gistic(r.getString(r.fieldIndex("Gene Symbol")), r.getInt(r.fieldIndex("Gene ID")), 
-          r.getString(r.fieldIndex("Cytoband")), columns.value.map(c => Sample(c, r.getInt(r.fieldIndex(c)))).toSeq
+        Gistic(r.getString(r.fieldIndex("Gene Symbol")), r.getLong(r.fieldIndex("Gene ID")), 
+          r.getString(r.fieldIndex("Cytoband")), columns.value.map(c => Sample(c, r.getLong(r.fieldIndex(c)))).toSeq
         )
     }}.as[Gistic]
   }
